+++ Cpp-C/Eta/Impl/Converter/jsmn.C
@@ -146,7 +146,7 @@ jsmnerr_t jsmn_parse(jsmn_parser *parser, const char *js, unsigned int len, jsmn
 	parser->len = len;
 
 	/* initialize the rest of tokens (they could be reallocated) */
-	for (i = parser->toknext; i < num_tokens; i++) {
+	for (i = parser->toknext; i < (int)num_tokens; i++) {
 		jsmn_fill_token(&tokens[i], JSMN_PRIMITIVE, -1, -1);
 	}
